split lib & main
